chore(root): Deploy all affected preview packages at once
See https://github.com/stackblitz-labs/pkg.pr.new/issues/247#issuecomment-2376044071
