+++ package.json
@@ -55,14 +55,14 @@
     "build:widget": "nx build @novu/widget",
     "build:embed": "nx build @novu/embed",
     "build:storybook": "nx run @novu/design-system:build-storybook",
-    "build:affected": "nx affected -t build --base=origin/next --head=HEAD --exclude=nextjs,@novu/shared",
     "test:providers": "cross-env pnpm --filter './packages/providers/**' test",
     "release:patch": "nx release version patch --no-push",
     "release:minor": "nx release version minor --no-push",
     "release:preminor": "nx release version preminor --no-push",
     "release:prerelease": "nx release version prerelease --no-push",
     "release:submodules": "ts-node scripts/release-with-submodules.ts",
-    "release:previews": "nx affected -t release:preview --base=origin/next --head=HEAD --exclude=nextjs,@novu/shared",
+    "preview:pkg:build": "nx affected -t build --base=origin/next --head=HEAD --exclude=nextjs",
+    "preview:pkg:publish": "node scripts/publish-preview-packages.mjs",
     "start:integration:api": "cd apps/api && pnpm run test",
     "start:e2e:api": "cd apps/api && pnpm run test:e2e",
     "g:module": "hygen module new --name=$npm_config_name",
