+++ packages/framework/package.json
@@ -29,8 +29,7 @@
     "$comment:bump:prerelease": "This is a workaround to support `npm version prerelease` with lerna",
     "bump:prerelease": "npm version prerelease --preid=alpha & PID=$!; (sleep 1 && kill -9 $PID) & wait $PID",
     "release:alpha": "pnpm bump:prerelease || pnpm build && npm publish",
-    "devtool": "tsx ./scripts/devtool.ts",
-    "release:preview": "pnpx pkg-pr-new publish"
+    "devtool": "tsx ./scripts/devtool.ts"
   },
   "keywords": [
     "novu",
