+++ packages/nest/package.json
@@ -26,8 +26,7 @@
     "watch:build": "tsc -p tsconfig.json -w",
     "watch:test": "jest src --watch",
     "reset-hard": "git clean -dfx && git reset --hard && pnpm install",
-    "prepare-release": "run-s reset-hard test",
-    "release:preview": "pnpx pkg-pr-new publish"
+    "prepare-release": "run-s reset-hard test"
   },
   "publishConfig": {
     "access": "public"
